--- a/lib/util.js
+++ b/lib/util.js
@@ -5,16 +5,16 @@ if (!Array.isArray) {
 }
 
 // deep clone an array of coordinates / polygons
-exports.clone = function(array) {
-  var new_arr = [];
-  for (var i = 0; i < array.length; i++) {
-    if (Array.isArray(array[i])){
-      new_arr.push(exports.clone(array[i]).slice());
-    } else {
-      new_arr.push(array[i]);
+exports.clone = function cloneArray(array) {
+  var newArray = array.slice();
+
+  for (var i = 0; i < newArray.length; i++) {
+    if (Array.isArray(newArray[i])) {
+      newArray[i] = cloneArray(newArray[i]);
     }
   }
-  return new_arr;
+
+  return newArray;
 }
 
 // Wrap an array of geometries to an array of polygons