Just a minor optimization. It looks like previously it was doing cloning twice on nested arrays.

```
# before
intersect#simple x 57,582 ops/sec ±0.98% (96 runs sampled)
intersect#armenia x 28,983 ops/sec ±0.83% (95 runs sampled)

# after
intersect#simple x 61,887 ops/sec ±0.83% (98 runs sampled)
intersect#armenia x 32,501 ops/sec ±0.78% (94 runs sampled)
```