@tchannel 2015-05-11
ah, forgot to turn travis on for the repo. It should be on now, but won't be triggered unless there's a new push.