@mourner 2015-05-11
Tests pass here, pinky-promise :)